+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 * Added support for the [Trunk](https://trunkrs.dev) wasm app build tool
+
+### Changed
 * `resolver` key is no longer cleaned from Cargo.toml
 
 ### [0.12.1] - 2023-04-10